# improve perf for default_signer: Don't create new PKey::RSA if we already can sign